API: Rename algorithms modules to something more useful
xdesign.algorithms and xdesign.geometry.algorithms are renamed to recon
and intersect respectively. These names are better because they tell
what the modules actually do. This module name is more specific and
hints at what is actually inside the module.
