+++ docs/source/api.rst
@@ -6,7 +6,6 @@ API
 .. toctree::
 
    api/xdesign.acquisition
-   api/xdesign.algorithms
    api/xdesign.geometry
    api/xdesign.constants
    api/xdesign.formats
@@ -14,6 +13,7 @@ API
    api/xdesign.metrics
    api/xdesign.phantom
    api/xdesign.plot
+   api/xdesign.recon
 
 .. automodule:: xdesign
    :members:
