+++ src/xdesign/geometry/intersect.py
